% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies-methods.R
\name{get_citations}
\alias{get_citations}
\title{Get citations for a set of study ids}
\usage{
get_citations(study_ids)
}
\arguments{
\item{study_ids}{is a vector of study ids}
}
\description{
Get citations for a set of study ids
}
