% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/studies.R
\name{get_study}
\alias{get_study}
\title{returns study for given ID}
\usage{
get_study(study_id = NULL, object_format = c("phylo", "nexml"),
  file_format = NULL, file)
}
\arguments{
\item{study_id}{}

\item{object_format}{}

\item{file_format}{}

\item{file}{}
}
\value{
some study
}
\description{
study
}
\examples{
\dontrun{
that_one_study <- get_study(study_id="pg_719", object_format="phylo")
}
}
\author{
Francois Michonneau
}

