<<<<<<< HEAD
% Generated by roxygen2 (4.0.2): do not edit by hand
=======
% Generated by roxygen2 (4.0.0): do not edit by hand
>>>>>>> exports sudies-family of functions, re-wrote docs for all
\name{get_study_subtree}
\alias{get_study_subtree}
\title{Retrieve subtree from a specific tree in the Open Tree of Life data store}
\usage{
get_study_subtree(study, tree, subtree_id)
}
\arguments{
\item{study}{character, study id}

\item{tree}{character, tree_id}

\item{subtree_id,}{either a node id that specifies a subtree or "ingroup"
which returns the ingroup is for this subtree, a 400 error otherwise}
}
\description{
Retrieve subtree from a specific tree in the Open Tree of Life data store
}
\examples{
small_tr <- get_study_subtree(study="pg_1144", tree="tree2324", subtree_id="node552052")
<<<<<<< HEAD
ingroup  <- get_study_subtree(study="pg_1144", tree="tree2324", subtree_id="ingroup")
=======
>>>>>>> exports sudies-family of functions, re-wrote docs for all
}

