% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_study_tree}
\alias{get_study_tree}
\title{Study Tree}
\usage{
get_study_tree(study = "pg_1144", tree = "tree2324", ...)
}
\arguments{
\item{study}{char study id}

\item{format}{char Tree format (default = json)}
}
\value{
A tree in desired format
}
\description{
returns specific treee from a study
}

