% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gol_node_info}
\alias{gol_node_info}
\title{Get summary information about a node in the graph}
\usage{
gol_node_info(node_id = NULL, ott_id = NULL, include_lineage = FALSE)
}
\arguments{
\item{node_id}{The idenitifer of the node in the graph.}

\item{ott_id}{The OpenTree taxonomic identifier.}

\item{include_lineage}{Boolean. Whether to return the lineage of the node from the synthetic tree.
Default = FALSE.}
}
\value{
A list of information about the node
}
\description{
Summary information about a queried node, including 1) whether it is in the graph DB,
2) whether it is in the synthetic tree, 3) supporting study sources, 4) number of
descendant tip taxa.
}
\details{
node info
}
\examples{
req <- gol_node_info(ott_id=81461)
}

