% Generated by roxygen2 (4.0.0): do not edit by hand
\name{gol_source_tree}
\alias{gol_source_tree}
\title{Get reconstructed source tree}
\usage{
gol_source_tree(study_id = NULL, tree_id = NULL, git_sha = NULL)
}
\arguments{
\item{study_id}{The study identifier. Will typically include a prefix ("pg_" or "ot_").}

\item{tree_id}{The tree identifier for a given study.}

\item{git_sha}{The git SHA identifying a particular source version.}

\item{format}{The name of the return format. The only currently supported format is newick.}
}
\value{
a newick tree
}
\description{
Returns a reconstructed source tree
}
\details{
source tree
}
\examples{
req <- gol_source_tree(study_id="pg_420", tree_id="522", git_sha="a2c48df995ddc9fd208986c3d4225112550c8452")
}

