% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/match_names.R
\name{list_synonyms_match_names}
\alias{list_synonyms_match_names}
\title{List the synonyms for a given name}
\usage{
list_synonyms_match_names(response, row_number, taxon_name, ott_id)
}
\arguments{
\item{response}{a data frame generated by}

\item{row_number}{}

\item{taxon_name}{}

\item{ott_id}{}
}
\value{
something
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Francois Michonneau
}

