% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/studies-methods.R
\name{list_trees}
\alias{list_trees}
\alias{list_trees.matched_studies}
\title{List trees ids in objects returned by
\code{\link{studies_find_studies}} and
\code{\link{studies_find_trees}}.}
\usage{
list_trees(matched_studies, ...)

\method{list_trees}{matched_studies}(matched_studies, study_id, ...)
}
\arguments{
\item{matched_studies}{an object created by
\code{studies_find_trees}}

\item{...}{Currently unused}

\item{study_id}{a \code{study_id} listed in the object returned by
\code{studies_find_trees}}
}
\value{
\code{list_trees} returns a list of the tree_ids for each
    study that match the requested criteria. If a \code{study_id}
    is provided, then only the trees for this study are returned
    as a vector.
}
\description{
List trees ids in objects returned by
\code{\link{studies_find_studies}} and
\code{\link{studies_find_trees}}.
}
\seealso{
\code{\link{studies_find_studies}} and
    \code{\link{studies_find_trees}}. The help for these functions
    have examples demonstrating the use of \code{list_trees}.
}

