% Generated by roxygen2 (4.0.2): do not edit by hand
\name{otl_about}
\alias{otl_about}
\title{Information about the tree of life}
\usage{
otl_about()
}
\value{
Some JSON
}
\description{
Basic information about the tree
}
\details{
Summary information about the current draft tree of life,
including information about the list of trees and the taxonomy
used to build it.
}
\author{
Francois Michonneau
}

