% Generated by roxygen2 (4.0.2): do not edit by hand
\name{otl_mrca}
\alias{otl_mrca}
\title{get MRCA}
\usage{
otl_mrca(q = list(ott_ids = c(412129, 536234)))
}
\arguments{
\item{nds}{}
}
\value{
the MRCA
}
\description{
Reurns the MRCA
}
\details{
Most recent common ancestor
}
\author{
Francois Michonneau
}

