% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/tol.R
\name{source_list}
\alias{source_list}
\alias{source_list.tol_summary}
\title{List of studies used in the Tree of Life}
\usage{
source_list(tax, ...)

\method{source_list}{tol_summary}(tax, ...)
}
\arguments{
\item{tax}{a list containing a \code{source_id_map} slot.}

\item{...}{additional arguments (currently unused)}
}
\value{
a data frame
}
\description{
Retrieve the detailed information for the list of studies used in
the Tree of Life.
}
\details{
This function takes the object resulting from
    \code{tol_about(study_list = TRUE)}, \code{tol_mrca()},
    \code{tol_node_info()}, and returns a data frame listing the
    \code{tree_id}, \code{study_id} and \code{git_sha} for the
    studies currently included in the Tree of Life.
}
