% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/studies.R
\name{studies_find_trees}
\alias{studies_find_trees}
\title{Find Trees}
\usage{
studies_find_trees(property = NULL, value = NULL, verbose = FALSE,
  exact = FALSE, ...)
}
\arguments{
\item{property}{The property to be searched on (character)}

\item{value}{The property-value to be searched on (character)}

\item{verbose}{Should the output include all metadata? (logical,
default \code{FALSE})}

\item{exact}{Should exact matching be used? (logical, default
\code{FALSE})}

\item{...}{additional arguments to customize the API request (see
\code{\link{rotl}} package documentation).}
}
\value{
A data frame that summarizes the trees found (and their
  associated studies) for the requested criteria. If a study has
  more than 5 trees, the \code{tree_ids} of the first ones will be
  shown followed by \code{...} to indicate that more are present.
}
\description{
Return a list of trees that match a given set of properties
}
\details{
The list of possible values to be used as values for the argument
\code{property} can be found using the function
\code{\link{studies_properties}}.
}
\examples{
\dontrun{
res <- studies_find_trees(property="ot:ottTaxonName", value="Drosophilia")
## summary of the trees and associated studies that match this criterion
res
## the full list of trees for each study
list_trees(res)
## the trees for a given study
list_trees(res, study_id = "pg_2769")
}
}
\seealso{
\code{\link{studies_properties}} which lists properties
  the studies can be searched on. \code{\link{list_trees}} for
  listing the trees that match the query.
}

