% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/studies.R
\name{studies_properties}
\alias{studies_properties}
\title{studies properties}
\usage{
studies_properties()
}
\value{
something
}
\description{
Property of a study
}
\examples{
all_the_properties <- studies_properties()
 unlist(all_the_properties$tree_properties)
}
\author{
Francois Michonneau
}

