% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tol.R
\name{study_list}
\alias{study_list}
\alias{study_list.tol_summary}
\title{List of studies used for the Tree of Life}
\usage{
study_list(tol)

\method{study_list}{tol_summary}(tol)
}
\arguments{
\item{tol}{an object created using \code{tol_about(study_list = TRUE)}}
}
\value{
a data frame
}
\description{
Retrieve the detailed information for the list of studies used in
the Tree of Life.
}
\details{
This function takes the object resulting from
\code{tol_about(study_list = TRUE)} and returns a data frame
listing the \code{tree_id}, \code{study_id} and \code{git_sha} for
the studies currently included in the Tree of Life.
}

