% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/taxonomy.R
\name{tax_lineage}
\alias{tax_lineage}
\alias{tax_lineage.taxon_info}
\title{Lineage of a taxon}
\usage{
tax_lineage(tax, ...)

\method{tax_lineage}{taxon_info}(tax, ...)
}
\arguments{
\item{tax}{an object created by \code{\link{taxonomy_taxon_info}}.}

\item{...}{additional arguments (currently unused).}
}
\value{
A list with one slot per taxon that contains a data frame
    with 3 columns: the taxonomy rank, the name, and unique name
    for all taxa included in the lineage of the taxon up to the
    root of the tree.
}
\description{
Extract the lineage information (higher taxonomy) from an object
returned by \code{\link{taxonomy_taxon_info}}.
}
\details{
The object passed to this function must have been created using
the argument \code{include_lineage=TRUE}.
}

