% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{tax_rank}
\alias{node_id}
\alias{ott_id}
\alias{ott_taxon_name}
\alias{synonyms}
\alias{tax_rank}
\title{Methods for Taxonomy}
\usage{
tax_rank(tax)

ott_taxon_name(tax)

node_id(tax)

ott_id(tax)

synonyms(tax)
}
\arguments{
\item{tax}{an object returned by \code{\link{taxonomy_taxon}} or
\code{\link{taxonomy_lica}}.}
}
\description{
Methods for dealing with objects returned by functions dealing
with the Taxonomy API.
}
\details{
This the page for the generic methods. See the help pages for
\code{\link{taxonomy_taxon}} and \code{\link{taxonomy_lica}} for
more information.
}

