% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy_about}
\alias{taxonomy_about}
\title{Taxonomy about}
\usage{
taxonomy_about()
}
\value{
Some JSON
}
\description{
Summary information about the OpenTree Taxaonomy (OTT)
}
\details{
Return information about the taxonomy, including version.
}
\examples{
taxonomy_about()
}

