% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy_lica}
\alias{taxonomy_lica}
\title{lica}
\usage{
taxonomy_lica(ott_ids = NULL)
}
\arguments{
\item{ott_ids}{The vector of ott ids for the taxa whose LICA is to be found. Not optional.}
}
\value{
Some JSON
}
\description{
Taxonomic LICA
}
\details{
Given a set of ott ids, get the taxon that is the least inclusive common ancestor
(the LICA) of all the identified taxa. A taxonomic LICA is analogous to a most recent
common ancestor (MRCA) in a phylogenetic tree.
}
\examples{
req <- taxonomy_lica(ott_ids=c(515698,590452,409712,643717))
}

