% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy_subtree}
\alias{taxonomy_subtree}
\title{taxon}
\usage{
taxonomy_subtree(ott_id = NULL, output_format = c("taxa_all", "newick",
  "phylo", "raw", "taxa_species", "taxa_internal"), file)
}
\arguments{
\item{ott_id}{The OTT id of the taxon of interest.  Not optional.}

\item{output_format}{}

\item{file}{}
}
\value{
Some JSON
}
\description{
Taxonomic Subtree
}
\details{
Given an ott id, return complete taxonomy subtree descended from specified taxon.
}
\examples{
req <- taxonomy_subtree(ott_id=515698)
}

