% Generated by roxygen2 (4.0.0): do not edit by hand
\name{taxonomy_taxon}
\alias{taxonomy_taxon}
\title{taxon}
\usage{
taxonomy_taxon(ott_id = NULL)
}
\arguments{
\item{ott_id}{The OTT id of the taxon of interest. Not optional.}
}
\value{
Some JSON
}
\description{
Taxon Info
}
\details{
Given an ott id, return information about the specified taxon.
}
\examples{
req <- taxonomy_taxon(ott_id=515698)
}

