% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tnrs.R
\name{tnrs_contexts}
\alias{tnrs_contexts}
\title{tnrs context}
\usage{
tnrs_contexts()
}
\value{
something
}
\description{
Return a list of pre-defined taxonomic contexts (i.e. clades),
which can be used to limit the scope of tnrs queries.
}
\details{
Taxonomic contexts are available to limit the scope of TNRS
searches. These contexts correspond to uncontested higher taxa
such as 'Animals' or 'Land plants'. This service returns a list
containing all available taxonomic context names, which may be
used as input (via the context_name parameter) to limit the search
scope of other services including match_names and
autocomplete_name.
}
\author{
Francois Michonneau
}

