% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tnrs.R
\name{tnrs_infer_context}
\alias{tnrs_infer_context}
\title{Infer context}
\usage{
tnrs_infer_context(names = NULL)
}
\arguments{
\item{names}{Vector of taxon names.}
}
\value{
something
}
\description{
Return a taxonomic context given a list of taxonomic names
}
\details{
Find the least inclusive taxonomic context that includes all the
unambiguous names in the input set. Unambiguous names are names
with exact matches to non-homonym taxa. Ambiguous names (those
without exact matches to non-homonym taxa) are indicated in
results.
}
\examples{
res <- tnrs_infer_context(names=c("Stellula calliope", "Struthio camelus"))
}
\author{
Francois Michonneau
}

