% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tnrs.R
\name{tnrs_match_names}
\alias{tnrs_match_names}
\title{Match names}
\usage{
tnrs_match_names(names = NULL, context_name = NULL,
  do_approximate_matching = TRUE, ids = NULL, include_deprecated = FALSE,
  include_dubious = FALSE)
}
\arguments{
\item{names}{taxon names to be queried (character vector)}

\item{context_name}{name of the taxonomic context to be searched
(length-one character vector)}

\item{do_approximate_matching}{A boolean indicating whether or not
to perform approximate string (a.k.a. "fuzzy") matching. Will
greatly improve speed if this is turned OFF (false). By
default, however, it is on (true).}

\item{ids}{An array of ids to use for identifying names. These
will be assigned to each name in the names array. If ids is
provided, then ids and names must be identical in length.}

\item{include_deprecated}{A boolean indicating whether or not to
include deprecated taxa in the search.}

\item{include_dubious}{Whether to include so-called 'dubious'
taxa--those which are not accepted by OTT.}
}
\value{
something
}
\description{
Match names
}
\details{
Accepts one or more taxonomic names and returns information about
potential matches for these names to known taxa in OTT. This
service uses taxonomic contexts to disambiguate homonyms and
misspelled names; a context may be specified using the
context_name parameter. If no context is specified, then the
context will be inferred: the shallowest taxonomic context that
contains all unambiguous names in the input set will be used. A
name is considered unambiguous if it is not a synonym and has only
one exact match to any taxon name in the entire taxonomy.
Taxonomic contexts are uncontested higher taxa that have been
selected to allow limits to be applied to the scope of TNRS
searches (e.g. 'match names only within flowering plants'). Once a
context has been identified (either user-specified or inferred),
all taxon name matches will performed only against taxa within
that context. For a list of available taxonomic contexts, see the
contexts service.
}
\examples{
\dontrun{
  deuterostomes <- tnrs_match_names(names=c("echinodermata", "xenacoelomorpha", "chordata", "hemichordata"))
}
}
\author{
Francois Michonneau
}

