% Generated by roxygen2 (4.0.0): do not edit by hand
\name{get_study}
\alias{get_study}
\title{study}
\usage{
get_study(study_id = NULL, object_format = c("phylo"), text_format = c("",
  "nexus", "newick", "json"), file)
}
\arguments{
\item{study}{}
}
\value{
some study
}
\description{
returns study for given ID
}
\details{
study
}
\examples{
\dontrun{
that_one_study <- get_study(study_id="pg_719", obj_format="phylo")
}
}
\author{
Francois Michonneau
}

