% Generated by roxygen2 (4.0.0): do not edit by hand
\name{get_study}
\alias{get_study}
\title{study}
\usage{
get_study(study)
}
\arguments{
\item{study}{}
}
\value{
some study
}
\description{
returns study for given ID
}
\details{
study
}
\examples{
that_one_study <- get_study(study="pg_719")
}
\author{
Francois Michonneau
}

