% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/studies.R
\name{get_study_meta}
\alias{get_study_meta}
\title{Study Metadata}
\usage{
get_study_meta(study_id, ...)
}
\arguments{
\item{study_id}{the study identifier (character)}

\item{...}{additional arguments to customize the API request (see
\code{\link{rotl}} package documentation).}
}
\value{
named-list containing the metadata associated with the
    study requested
}
\description{
Retrieve metadata about a study in the Open Tree of Life datastore
}
\examples{
req <- get_study_meta("pg_719")
req$nexml$`^ot:studyPublication`
}

