% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_study_meta}
\alias{get_study_meta}
\title{Study Metadata}
\usage{
get_study_meta(study_id)
}
\arguments{
\item{study}{character, study id}
}
\value{
named-list json file with metadata
}
\description{
Retrieve metadata about a study in the Open Tree of Life datastore
}
\examples{
req <- get_study_meta("pg_719")
req$nexml$`^ot:studyPublication`
}

