% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/studies.R
\name{get_study_subtree}
\alias{get_study_subtree}
\title{Retrieve subtree from a specific tree in the Open Tree of Life data store}
\usage{
get_study_subtree(study_id, tree_id, subtree_id, object_format = c("phylo"),
  file_format = NULL, file)
}
\arguments{
\item{study_id}{character, the study id}

\item{tree_id}{character,  the tree id}

\item{subtree_id,}{either a node id that specifies a subtree or "ingroup"
which returns the ingroup is for this subtree, a 400 error otherwise}

\item{object_format}{character, the class of the object returned
by the function (default, and currently only possibility
\code{phylo} from the APE package)}

\item{file_format}{character, the file format to use to save the
results of the query.}

\item{file}{character, the path and file name where the output
should be written.}
}
\description{
Retrieve subtree from a specific tree in the Open Tree of Life data store
}
\examples{
\dontrun{
small_tr <- get_study_subtree(study_id="pg_1144", tree="tree2324", subtree_id="node552052")
ingroup  <- get_study_subtree(study_id="pg_1144", tree="tree2324", subtree_id="ingroup")
}
}

