% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/studies.R
\name{get_study_tree}
\alias{get_study_tree}
\title{Study Tree}
\usage{
get_study_tree(study_id = NULL, tree_id = NULL,
  object_format = c("phylo"), tip_label = c("original_label", "ott_id",
  "ott_taxon_name"), file_format = NULL, file)
}
\arguments{
\item{study_id}{}

\item{tree_id}{}

\item{object_format}{}

\item{tip_label}{}

\item{file_format}{}

\item{file}{}

\item{format}{char Tree format (default = json)}
}
\value{
A tree file in desired format
}
\description{
returns specific tree from a study
}
\examples{
\dontrun{
 nexson_tr <- get_study_tree(study_id="pg_1144", tree="tree2324")
}
}

