<<<<<<< HEAD
% Generated by roxygen2 (4.0.2): do not edit by hand
=======
% Generated by roxygen2 (4.0.0): do not edit by hand
>>>>>>> exports sudies-family of functions, re-wrote docs for all
\name{get_study_tree}
\alias{get_study_tree}
\title{Study Tree}
\usage{
get_study_tree(study, tree, format)
}
\arguments{
\item{study}{char study id}

<<<<<<< HEAD
\item{tree}{tree id}

=======
>>>>>>> exports sudies-family of functions, re-wrote docs for all
\item{format}{char Tree format (default = json)}
}
\value{
A tree file in desired format
}
\description{
returns specific tree from a study
}
\examples{
nexson_tr <- get_study_tree(study="pg_1144", tree="tree2324")
}

