% Generated by roxygen2 (4.0.0): do not edit by hand
\name{get_study_tree}
\alias{get_study_tree}
\title{Study Tree}
\usage{
get_study_tree(study_id = NULL, tree_id = NULL,
  object_format = c("phylo"), text_format = NULL, file)
}
\arguments{
\item{study}{char study id}

\item{tree}{tree id}

\item{format}{char Tree format (default = json)}
}
\value{
A tree file in desired format
}
\description{
returns specific tree from a study
}
\examples{
nexson_tr <- get_study_tree(study_id="pg_1144", tree="tree2324")
}

