% Generated by roxygen2 (4.0.0): do not edit by hand
\name{gol_source_tree}
\alias{gol_source_tree}
\title{Get reconstructed source tree}
\usage{
gol_source_tree(study_id = NULL, tree_id = NULL, git_sha = NULL)
}
\arguments{
\item{study_id}{String. The study identifier. Will typically include a prefix ("pg_" or "ot_").}

\item{tree_id}{String. The tree identifier for a given study.}

\item{git_sha}{String. The git SHA identifying a particular source version.}

\item{format}{The name of the return format. The only currently supported format is newick.}
}
\value{
a tree of class \code{"phylo"}
}
\description{
Returns a reconstructed source tree from the graph DB.
}
\details{
Reconstructs a source tree given identifiers: \code{study_id}, \code{tree_id}, and \code{git_sha}. The tree may differ from the original source tree in 2 ways: 1) it may contain fewer taxa (as duplicate taxa are pruned on tree ingestion), and 2) OpenTree Taxonomy IDs (ottIDs) are applied to all named internal nodes and as a suffix to all terminal node names.
}
\examples{
res <- gol_source_tree(study_id="ot_121", git_sha="a2c48df995ddc9fd208986c3d4225112550c8452", tree_id="7")
}

