% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/match_names.R
\name{list_synonyms_match_names}
\alias{list_synonyms_match_names}
\title{List the synonyms for a given name}
\usage{
list_synonyms_match_names(response, row_number, taxon_name, ott_id)
}
\arguments{
\item{response}{a data frame generated by the tnrs_match_names
function}

\item{row_number}{the row number corresponding to the name for
which to list the synonyms}

\item{taxon_name}{the taxon name corresponding to the name for
which to list the synonyms}

\item{ott_id}{the ott id corresponding to the name for which to
list the synonyms}
}
\value{
a list whose elements are all synomym names (as vectors of
    character) for the taxonomic names that match the query (the
    names of the elements of the list).
}
\description{
When querying the Taxonomic Name Resolution Services for a
particular taxonomic name, the API returns as possible matches all
names that include the queried name as a possible synonym. This
function allows you to explore other synonyms for an accepted
name, and allows you to determine why the name you queried is
returning an accepted synonym.
}
\details{
To list synonyms for a given taxonomic name, you need to provide
the object resulting from a call to the tnrs_match_names function,
as well as one of either the row number corresponding to the name
in this object, the name itself (as used in the original query),
or the ott_id listed for this name.
}

