% Generated by roxygen2 (4.0.2): do not edit by hand
\name{studies_find_studies}
\alias{studies_find_studies}
\title{find_study}
\usage{
studies_find_studies(property = NULL, value = NULL, verbose = FALSE,
  exact = FALSE)
}
\arguments{
\item{property}{character, the property to be searched on}

\item{value}{character, the property-value  to be searched on}

\item{verbose}{Boolean, include all metadata (default=FALSE)}

\item{exact}{Boolean, exact matching (default = FALSE)}
}
\description{
Return a list of studes that match a given properties
}
\examples{
all_s <- studies_find_studies()
length(all_s$
}

