% Generated by roxygen2 (4.0.0): do not edit by hand
\name{studies_find_studies}
\alias{studies_find_studies}
\title{find_study}
\usage{
studies_find_studies(property = NULL, value = NULL, verbose = FALSE,
  exact = FALSE)
}
\arguments{
\item{exact}{Boolean, exact matching (default = FALSE)}

\item{property}{character, the property to be searched on}

\item{value}{character, the property-value  to be searched on}

\item{verbose}{Boolean, include all metadata (default=FALSE)}
}
\description{
Return a list of studies that match a given properties
}
\examples{
req <- studies_find_studies(property="ot:studyId", value="pg_719")
httr::content(req)
}
\seealso{
\code{\link{studies_properties}} which lists properties against
which the studies can be searched
}

