% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/studies.R
\name{studies_find_studies}
\alias{studies_find_studies}
\title{find_study}
\usage{
studies_find_studies(property = NULL, value = NULL, verbose = FALSE,
  exact = FALSE, ...)
}
\arguments{
\item{property}{The property to be searched on (character)}

\item{value}{The property-value to be searched on (character)}

\item{verbose}{Should the output include all metadata (logical
default \code{FALSE})}

\item{exact}{Should exact matching be used? (logical, default
\code{FALSE})}

\item{...}{additional arguments to customize the API request (see
\code{\link{rotl}} package documentation).}
}
\description{
Return a list of studies that match given properties
}
\examples{
\dontrun{
study <- studies_find_studies(property="ot:studyId", value="pg_719")
}
}
\seealso{
\code{\link{studies_properties}} which lists properties
    against which the studies can be searched
}

