% Generated by roxygen2 (4.0.0): do not edit by hand
\name{studies_find_trees}
\alias{studies_find_trees}
\title{find trees}
\usage{
studies_find_trees(property = NULL, value = NULL, verbose = FALSE,
  exact = FALSE)
}
\arguments{
\item{property}{character, the property to be searched on}

\item{value}{character, the property-value  to be searched on}

\item{verbose}{Boolean, include all metadata (default=FALSE)}

\item{exact}{Boolean, exact matching (default = FALSE)}
}
\description{
Return a list of trees that match a given properties
}
\examples{
req <- studies_find_trees(property="ot:ottTaxonName", value="Garcinia")
length(httr::content(req)$matched_studies)
}
\seealso{
\code{\link{studies_properties}} which lists properties against
which the studies can be searched
}
\seealso{
\code{\link{studies_properties}} which lists properties against
which the studies can be searched
}

