% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/studies.R
\name{studies_find_trees}
\alias{studies_find_trees}
\title{find trees}
\usage{
studies_find_trees(property = NULL, value = NULL, verbose = FALSE,
  exact = FALSE, ...)
}
\arguments{
\item{property}{The property to be searched on (character)}

\item{value}{The property-value to be searched on (character)}

\item{verbose}{Should the output include all metadata? (logical,
default \code{FALSE})}

\item{exact}{Should exact matching be used? (logical, default
\code{FALSE})}

\item{...}{additional arguments to customize the API request (see
\code{\link{rotl}} package documentation).}
}
\description{
Return a list of trees that match a given properties
}
\details{
The list of possible values to be used as values for the argument
\code{property} can be found using the function
\code{\link{studies_properties}}.
}
\examples{
\dontrun{
res <- studies_find_trees(property="ot:ottTaxonName", value="Garcinia")
}
}
\seealso{
\code{\link{studies_properties}} which lists properties
    against which the studies can be searched
}

