% Generated by roxygen2 (4.0.0): do not edit by hand
\name{studies_properties}
\alias{studies_properties}
\title{studies properties}
\usage{
studies_properties()
}
\value{
<<<<<<< HEAD
something
=======
httr request containing lists of properties
>>>>>>> exports sudies-family of functions, re-wrote docs for all
}
\description{
returns properties on which studies and study-trees can be searched
Property of a study
}
\details{
properties of the studies
}
\examples{
all_the_properties <- studies_properties()
 prop_list = httr::content(all_the_properties)
 unlist(prop_list$tree_properties)
}
\author{
Francois Michonneau
}
\examples{
all_the_properties <- studies_properties()
prop_list <- httr::content(all_the_properties)
unlist(prop_list$tree_properties)
}

