% Generated by roxygen2 (4.0.2): do not edit by hand
\name{studies_properties}
\alias{studies_properties}
\title{studies properties}
\usage{
studies_properties()
}
\value{
something
}
\description{
Property of a study
}
\examples{
all_the_properties <- studies_properties()
 prop_list = httr::content(all_the_properties)
 unlist(prop_list$tree_properties)
}
\author{
Francois Michonneau
}

