% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies.R
\name{studies_properties}
\alias{studies_properties}
\title{Properties of the Studies}
\usage{
studies_properties(...)
}
\arguments{
\item{...}{additional arguments to customize the API request (see
\code{\link{rotl}} package documentation).}
}
\value{
A list of the study properties that can be used to find
    studies and trees that are contributing to the synthetic tree.
}
\description{
Return the list of study properties that can be used to search
studies and trees used in the synthetic tree.
}
\details{
The list returned has 2 elements \code{tree_properties} and
\code{studies_properties}. Each of these elements lists
theadditional arguments to customize the API request properties
that can be used to search for trees and studies that are
contributing to the synthetic tree.
}
\examples{
\dontrun{
 all_the_properties <- studies_properties()
 unlist(all_the_properties$tree_properties)
}
}
\seealso{
\code{\link{studies_find_trees}}
}

