% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{tax_rank}
\alias{ott_id}
\alias{ott_taxon_name}
\alias{synonyms}
\alias{tax_rank}
\title{Methods for Taxonomy}
\usage{
tax_rank(tax)

ott_taxon_name(tax)

ott_id(tax, ...)

synonyms(tax, ...)
}
\arguments{
\item{tax}{an object returned by \code{\link{taxonomy_taxon}},
\code{\link{taxonomy_lica}}, or \code{\link{tnrs_match_names}}}

\item{...}{additional arguments (see
\code{\link{tnrs_match_names}})}
}
\description{
Methods for dealing with objects returned by functions dealing
with the Taxonomy and the Taxonomic Name Resolution Services APIs.
}
\details{
This is the page for the generic methods. See the help pages for
\code{\link{taxonomy_taxon}}, \code{\link{taxonomy_lica}}, and
\code{\link{tnrs_match_names}} for more information.
}

