% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tax_rank}
\alias{tax_rank}
\alias{ott_id}
\alias{synonyms}
\alias{tax_sources}
\alias{is_suppressed}
\alias{unique_name}
\alias{tax_name}
\title{Methods for Taxonomy}
\usage{
tax_rank(tax, ...)

ott_id(tax, ...)

synonyms(tax, ...)

tax_sources(tax, ...)

is_suppressed(tax, ...)

unique_name(tax, ...)

tax_name(tax, ...)
}
\arguments{
\item{tax}{an object returned by \code{\link{taxonomy_taxon_info}},
\code{\link{taxonomy_mrca}}, or \code{\link{tnrs_match_names}}}

\item{...}{additional arguments (see
\code{\link{tnrs_match_names}})}
}
\description{
Methods for dealing with objects containing taxonomic information
(Taxonomy, TNRS endpoints)
}
\details{
This is the page for the generic methods. See the help pages for
\code{\link{taxonomy_taxon_info}}, \code{\link{taxonomy_mrca}}, and
\code{\link{tnrs_match_names}} for more information.
}
