% Generated by roxygen2 (4.0.0): do not edit by hand
\name{taxonomy_about}
\alias{taxonomy_about}
\title{Taxonomy about}
\usage{
taxonomy_about()
}
\value{
Some JSON
}
\description{
Summary information about the OpenTree Taxaonomy (OTT)
}
\details{
Return information about the taxonomy, including version.
}

