% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy_about}
\alias{taxonomy_about}
\title{About the Open Tree Taxonomy}
\usage{
taxonomy_about(...)
}
\arguments{
\item{...}{additional arguments to customize the API request (see
\code{\link{rotl}} package documentation).}
}
\value{
A list with at least the version and the location of the
    taxonomy source files
}
\description{
Summary information about the Open Tree Taxaonomy (OTT)
}
\details{
Return metadata and information about the taxonomy
itself. Currently, the available metadata is fairly sparse, but
includes (at least) the eversion, and the location from which the
complete taxonomy source files can be downloaded.
}
\examples{
\dontrun{
taxonomy_about()
}
}

