% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy_about}
\alias{taxonomy_about}
\title{Information about the Open Tree Taxonomy}
\usage{
taxonomy_about(...)
}
\arguments{
\item{...}{additional arguments to customize the API request (see
\code{\link{rotl}} package documentation).}
}
\value{
A list with the following properties:
\itemize{

    \item {weburl} {String. The release page for this version
    of the taxonomy.}

    \item {author} {String. The author string.}

    \item {name} {String. The name of the taxonomy.}

    \item {source} {String. The full identifying information for
    this version of the taxonomy.}

    \item {version} {String. The version number of the taxonomy.}
}
}
\description{
Summary information about the Open Tree Taxaonomy (OTT)
}
\details{
Return metadata and information about the taxonomy
itself. Currently, the available metadata is fairly sparse, but
includes (at least) the version, and the location from which the
complete taxonomy source files can be downloaded.
}
\examples{
\dontrun{
taxonomy_about()
}
}

