% Generated by roxygen2 (4.0.0): do not edit by hand
\name{taxonomy_subtree}
\alias{taxonomy_subtree}
\title{taxon}
\usage{
taxonomy_subtree(ott_id = NULL)
}
\arguments{
\item{ott_id}{The OTT id of the taxon of interest.  Not optional.}
}
\value{
Some JSON
}
\description{
Taxonomic Subtree
}
\details{
Given an ott id, return complete taxonomy subtree descended from specified taxon.
}
\examples{
req <- taxonomy_subtree(ott_id=515698)
}

