% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tnrs_infer_context}
\alias{tnrs_infer_context}
\title{infer context}
\usage{
tnrs_infer_context(taxon_names = NULL)
}
\arguments{
\item{taxon_names}{}
}
\value{
something
}
\description{
Return a taxonomic context given a list of taxonomic names.
}
\details{
Find the least inclusive taxonomic context that includes all the
unambiguous names in the input set. Unambiguous names are names
with exact matches to non-homonym taxa. Ambiguous names (those
without exact matches to non-homonym taxa) are indicated in
results.
}
\author{
Francois Michonneau
}

