% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tol_about}
\alias{tol_about}
\title{Information about the tree of life}
\usage{
tol_about(study_list = FALSE)
}
\arguments{
\item{study_list}{Boolean. Whether to return the list of source studies. Default = FALSE.}
}
\value{
Some JSON
}
\description{
Basic information about the tree
}
\details{
Summary information about the current draft tree of life,
including information about the list of trees and the taxonomy
used to build it.
}
\author{
Francois Michonneau
}

