% Generated by roxygen2 (4.0.0): do not edit by hand
\name{tol_induced_subtree}
\alias{tol_induced_subtree}
\title{induced subtree}
\usage{
tol_induced_subtree(node_ids = NULL, ott_ids = NULL)
}
\arguments{
\item{node_ids}{Node ids indicating nodes to be used as tips in
the induced tree}

\item{ott_ids}{OTT ids indicating nodes to be used as tips in the
induced tree}
}
\value{
a tree of class \code{"phylo"}
}
\description{
Extract induced subtree
}
\details{
Return a tree with tips corresponding to the nodes identified in
the input set(s), that is consistent with topology of the current
draft tree. This tree is equivalent to the minimal subtree induced
on the draft tree by the set of identified nodes. Any combination
of node ids and ott ids may be used as input. Nodes or ott ids
that do not correspond to any found nodes in the graph, or which
are in the graph but are absent from the synthetic tree, will be
identified in the output (but will of course not be present in the
resulting induced tree). Branch lengths in the result may be
arbitrary, and the leaf labels of the tree may either be taxonomic
names or (for nodes not corresponding directly to named taxa) node
ids.
}
\examples{
res <- tol_induced_subtree(ott_ids=c(292466, 501678, 267845, 666104, 316878, 102710, 176458))
}
\author{
Francois Michonneau
}

