% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tol.R
\name{tol_mrca}
\alias{ott_id.tol_mrca}
\alias{source_list.tol_mrca}
\alias{tax_name.tol_mrca}
\alias{tax_rank.tol_mrca}
\alias{tax_sources.tol_mrca}
\alias{tol_mrca}
\alias{unique_name.tol_mrca}
\title{MRCA of taxa from the synthetic tree}
\usage{
tol_mrca(ott_ids = NULL, node_ids = NULL, ...)

\method{tax_sources}{tol_mrca}(tax, ...)

\method{unique_name}{tol_mrca}(tax, ...)

\method{tax_name}{tol_mrca}(tax, ...)

\method{tax_rank}{tol_mrca}(tax, ...)

\method{ott_id}{tol_mrca}(tax, ...)

\method{source_list}{tol_mrca}(tax, ...)
}
\arguments{
\item{ott_ids}{Numeric vector. The ott ids for which the MRCA is desired.}

\item{node_ids}{Character vector. The node ids for which the MRCA is desired.}

\item{...}{additional arguments to customize the API call (see
\code{\link{rotl}} for more information).}

\item{tax}{an object returned by \code{tol_mrca()}.}
}
\value{
A list
}
\description{
Most Recent Common Ancestor for a set of nodes
}
\examples{
\dontrun{
birds_mrca <- tol_mrca(ott_ids=c(412129, 536234))
ott_id(birds_mrca)
tax_sources(birds_mrca)
}
}

