% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tol.R
\name{tol_mrca}
\alias{tol_mrca}
\title{MRCA of taxa from the synthetic tree}
\usage{
tol_mrca(ott_ids = NULL, ...)
}
\arguments{
\item{ott_ids}{the ott ids for which the MRCA is desired
(character or numeric vector)}

\item{...}{additional arguments to customize the API call (see
\code{\link{rotl}} for more information).}
}
\value{
A list
}
\details{
Return the most recent common ancestor of a set of nodes
    in the synthetic tree.
}
\examples{
\dontrun{
birds_mrca <- tol_mrca(ott_ids=c(412129, 536234))
}
}

