% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tol.R
\name{tol_mrca}
\alias{tol_mrca}
\title{MRCA of taxa from the synthetic tree}
\usage{
tol_mrca(ott_ids = NULL, node_ids = NULL, ...)
}
\arguments{
\item{ott_ids}{Numeric vector. The ott ids for which the MRCA is desired.}

\item{node_ids}{Character vector. The node ids for which the MRCA is desired.}

\item{...}{additional arguments to customize the API call (see
\code{\link{rotl}} for more information).}
}
\value{
A list
}
\description{
Most Recent Common Ancestor for a set of nodes
}
\details{
Get the MRCA of a set of nodes on the current synthetic
tree. Accepts any combination of node ids and ott ids as
input. Returns information about the most recent common ancestor
(MRCA) node as well as the most recent taxonomic ancestor (MRTA)
node (the closest taxonomic node to the MRCA node in the synthetic
tree; the MRCA and MRTA may be the same node). If any of the
specified nodes is not in the synthetic tree an error will be returned.
}
\examples{
\dontrun{
birds_mrca <- tol_mrca(ott_ids=c(412129, 536234))
}
}

