% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tol_mrca}
\alias{tol_mrca}
\title{get MRCA}
\usage{
tol_mrca(ott_ids, node_ids)
}
\arguments{
\item{ott_ids}{}

\item{node_ids}{}
}
\value{
the MRCA
}
\description{
Reurns the MRCA
}
\details{
Most recent common ancestor
}
\examples{
test1 <- tol_mrca(ott_ids=list("ott_ids" = c(412129, 536234)))
test2 <- tol_mrca(ott_ids=list("ott_ids" = c(415255)), node_ids=c(341556))
}
\author{
Francois Michonneau
}

