% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tol_mrca}
\alias{tol_mrca}
\title{get MRCA}
\usage{
tol_mrca(ott_ids = NULL, node_ids = NULL)
}
\arguments{
\item{ott_ids}{}

\item{node_ids}{}
}
\value{
the MRCA
}
\description{
Most recent common ancestor of nodes
}
\details{
Return the most recent common ancestor of a set of nodes in the synthetic tree.
}
\examples{
test1 <- tol_mrca(ott_ids=c(412129, 536234))
}
\author{
Francois Michonneau
}

