% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tol.R
\name{tol_subtree}
\alias{tol_subtree}
\title{Extract subtree}
\usage{
tol_subtree(node_id = NULL, ott_id = NULL, tree_id = NULL)
}
\arguments{
\item{node_id}{The node id of the node in the tree that should
serve as the root of the tree returned. This argument may not be
used in combination with ott_id.}

\item{ott_id}{The ott id of the node in the tree that should serve
as the root of the tree returned. This argument may not be used in
combination with node_id.}

\item{tree_id}{The identifier for the synthesis tree. We currently
only support a single draft tree in the db at a time, so this
argument is superfluous and may be safely ignored.}
}
\value{
a tree of class \code{"phylo"}
}
\description{
Extract subtree from a node or ott id
}
\details{
Return a complete subtree of the draft tree descended from some
specified node. The node to use as the start node may be specified
using either a node id or an ott id, but not both. If the
specified node is not in the synthetic tree (or is entirely absent
from the graph), an error will be returned.
}
\examples{
\dontrun{
res <- tol_subtree(ott_id=81461)
}
}
\author{
Francois Michonneau
}

